+++ packages/components/.storybook/decorators/index.ts
@@ -1,2 +1,3 @@
 export { default as withTokenSetup } from './withTokenSetup'
 export { default as withProviders } from './withProviders'
+export { default as withMutationResolver } from './withMutationResolver'