+++ packages/components/modules/profiles/web/ProfileSettingsComponent/types.ts
@@ -1,0 +1,5 @@
+import { ProfileComponentFragment$key } from '../../../../__generated__/ProfileComponentFragment.graphql'
+
+export interface ProfileSettingsComponentProps {
+  profile?: ProfileComponentFragment$key | null
+}